azure-function-chat/requirements.txt
azure-functions==1.18.0
azure-search-documents==11.4.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.19.0
openai==1.3.8
python-dotenv==1.0.0
typing-extensions==4.8.0
aiohttp==3.9.0
requests==2.31.0
pydantic==2.5.0
tenacity==8.2.0
pandas==2.0.0
numpy==1.24.0